// SDKs/bitbucket.org/2.0/typescript/src/sdk/models/shared/paginatedbranchrestrictions.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


// PaginatedBranchrestrictions
/** 
 * A paginated list of branch restriction rules.
**/
export class PaginatedBranchrestrictions extends SpeakeasyBase {
  @Metadata({ data: "json, name=next" })
  next?: string;

  @Metadata({ data: "json, name=page" })
  page?: number;

  @Metadata({ data: "json, name=pagelen" })
  pagelen?: number;

  @Metadata({ data: "json, name=previous" })
  previous?: string;

  @Metadata({ data: "json, name=size" })
  size?: number;

  @Metadata({ data: "json, name=values" })
  values?: Map<string, any>[];
}