// SDKs/bitbucket.org/2.0/typescript/src/sdk/models/operations/getrepositoriesworkspacereposlugbranchrestrictions.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import * as shared from "../shared";


export class GetRepositoriesWorkspaceRepoSlugBranchRestrictionsPathParams extends SpeakeasyBase {
  @Metadata({ data: "pathParam, style=simple;explode=false;name=repo_slug" })
  repoSlug: string;

  @Metadata({ data: "pathParam, style=simple;explode=false;name=workspace" })
  workspace: string;
}


export class GetRepositoriesWorkspaceRepoSlugBranchRestrictionsQueryParams extends SpeakeasyBase {
  @Metadata({ data: "queryParam, style=form;explode=true;name=kind" })
  kind?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=pattern" })
  pattern?: string;
}


export class GetRepositoriesWorkspaceRepoSlugBranchRestrictionsSecurityOption1 extends SpeakeasyBase {
  @Metadata({ data: "security, scheme=true;type=oauth2" })
  oauth2: shared.SchemeOauth2;
}


export class GetRepositoriesWorkspaceRepoSlugBranchRestrictionsSecurityOption2 extends SpeakeasyBase {
  @Metadata({ data: "security, scheme=true;type=http;subtype=basic" })
  basic: shared.SchemeBasic;
}


export class GetRepositoriesWorkspaceRepoSlugBranchRestrictionsSecurityOption3 extends SpeakeasyBase {
  @Metadata({ data: "security, scheme=true;type=apiKey;subtype=header" })
  apiKey: shared.SchemeApiKey;
}


export class GetRepositoriesWorkspaceRepoSlugBranchRestrictionsSecurity extends SpeakeasyBase {
  @Metadata({ data: "security, option=true" })
  option1?: GetRepositoriesWorkspaceRepoSlugBranchRestrictionsSecurityOption1;

  @Metadata({ data: "security, option=true" })
  option2?: GetRepositoriesWorkspaceRepoSlugBranchRestrictionsSecurityOption2;

  @Metadata({ data: "security, option=true" })
  option3?: GetRepositoriesWorkspaceRepoSlugBranchRestrictionsSecurityOption3;
}


export class GetRepositoriesWorkspaceRepoSlugBranchRestrictionsRequest extends SpeakeasyBase {
  @Metadata()
  pathParams: GetRepositoriesWorkspaceRepoSlugBranchRestrictionsPathParams;

  @Metadata()
  queryParams: GetRepositoriesWorkspaceRepoSlugBranchRestrictionsQueryParams;

  @Metadata()
  security: GetRepositoriesWorkspaceRepoSlugBranchRestrictionsSecurity;
}


export class GetRepositoriesWorkspaceRepoSlugBranchRestrictionsResponse extends SpeakeasyBase {
  @Metadata()
  contentType: string;

  @Metadata()
  statusCode: number;

  @Metadata()
  error?: Map<string, any>;

  @Metadata()
  paginatedBranchrestrictions?: shared.PaginatedBranchrestrictions;
}
